# docs/api-docs/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "中间层模块",
        "description": "虚拟宠物医院API文档，前端用。字段默认值：int为-1，String为'This field is currently Empty.'，Date为null，应该会表现为空字符串。",
        "version": "1.0.1-SNAPSHOT"
    },
    "servers": [
        {
            "url": "http://pikapikpinoimac:8090",
            "description": "Generated server url"
        }
    ],
    "paths": {
...
        "/api/exam/question-in-paper": {
            "get": {
                "tags": [
                    "测验模块"
                ],
                "summary": "获取试卷中问题列表接口",
                "operationId": "getQuestionInPapers",
                "parameters": [
                    {
                        "name": "questionPoint",
                        "in": "query",
                        "description": "试卷中问题分数",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "paperId",
                        "in": "query",
                        "description": "试卷Id",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "questionId",
                        "in": "query",
                        "description": "问题Id",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": 200,
                                    "data": {
                                        "questionInPaperList": [
                                            {
                                                "questionInPaperPaper": {
                                                    "paperDuration": "2 hours",
                                                    "paperTotalScore": "100",
                                                    "paperExam": {
                                                        "examName": "Math Exam",
                                                        "examId": 1
                                                    },
                                                    "paperName": "Math Paper",
                                                    "paperId": 1
                                                },
                                                "questionPoint": 3,
                                                "questionInPaperQuestion": {
                                                    "questionContent": "test new",
                                                    "questionId": 2,
                                                    "questionCategory": {
                                                        "categoryName": "Algebra",
                                                        "categoryId": 1
                                                    },
                                                    "questionType": "qa"
                                                },
                                                "questionInPaperId": 2
                                            },
                                            {
                                                "questionInPaperPaper": {
                                                    "paperDuration": "2 hours",
                                                    "paperTotalScore": "100",
                                                    "paperExam": {
                                                        "examName": "Math Exam",
                                                        "examId": 1
                                                    },
                                                    "paperName": "Math Paper",
                                                    "paperId": 1
                                                },
                                                "questionPoint": 3,
                                                "questionInPaperQuestion": {
                                                    "questionContent": "new question 3",
                                                    "questionId": 8,
                                                    "questionCategory": {
                                                        "categoryName": "Algebra",
                                                        "categoryId": 1
                                                    },
                                                    "questionType": "type1"
                                                },
                                                "questionInPaperId": 3
                                            }
                                        ]
                                    },
                                    "message": "ok"
                                }
                            }
                        }
                    },
                    "515": {
                        "description": "Request failed",
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": 515,
                                    "message": "Request failed"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "测验模块"
                ],
                "summary": "更改试卷中问题信息接口",
                "operationId": "updateQuestionInPapers",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "json",
                                "description": "更改后的试卷中问题信息",
                                "example": {
                                    "questionInPaperId": 3,
                                    "questionPoint": 3,
                                    "paperId": 1,
                                    "questionId": 2
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "515": {
                        "description": "Request failed",
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": 515,
                                    "message": "Request failed"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": 200,
                                    "data": {
                                        "questionInPaperPaper": {
                                            "paperDuration": "2 hours",
                                            "paperTotalScore": "100",
                                            "paperExam": {
                                                "examName": "Math Exam",
                                                "examId": 1
                                            },
                                            "paperName": "Math Paper",
                                            "paperId": 1
                                        },
                                        "questionPoint": 8,
                                        "questionInPaperQuestion": {
                                            "questionContent": "new question 3",
                                            "questionId": 8,
                                            "questionCategory": {
                                                "categoryName": "Algebra",
                                                "categoryId": 1
                                            },
                                            "questionType": "type1"
                                        },
                                        "questionInPaperId": 3
                                    },
                                    "message": "ok"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "测验模块"
                ],
                "summary": "新增试卷中问题接口",
                "operationId": "addQuestionInPapers",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "json",
                                "description": "新增的试卷中问题信息",
                                "example": {
                                    "questionPoint": 3,
                                    "paperId": 1,
                                    "questionId": 2
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "515": {
                        "description": "Request failed",
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": 515,
                                    "message": "Request failed"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": 200,
                                    "data": {
                                        "questionInPaperPaper": {
                                            "paperDuration": "2 hours",
                                            "paperTotalScore": "100",
                                            "paperExam": {
                                                "examName": "Math Exam",
                                                "examId": 1
                                            },
                                            "paperName": "Math Paper",
                                            "paperId": 1
                                        },
                                        "questionPoint": 3,
                                        "questionInPaperQuestion": {
                                            "questionContent": "new question 3",
                                            "questionId": 8,
                                            "questionCategory": {
                                                "categoryName": "Algebra",
                                                "categoryId": 1
                                            },
                                            "questionType": "type1"
                                        },
                                        "questionInPaperId": 3
                                    },
                                    "message": "ok"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}